requirements.txt
opencv-python
numpy
matplotlib
ultralytics
easyocr
torch
torchvision
torchaudio
requests
tensorflow
Pillow
pymongo 